feat: 시도할 횟수를 입력 받을 때 예외 상황 추가

+++ src/main/java/racingcar/Application.java
@@ -40,10 +40,10 @@ public class Application {
 
 
         input = Console.readLine();
-        int count = 0;
+        long count = 0;
 
         try {
-            count = Integer.parseInt(input);
+            count = Long.parseLong(input);
         } catch (NumberFormatException ex) {
             throw new IllegalArgumentException();
         }
